Создал класс BaseTest, куда положил @After и @Before
