+++ src/test/java/annotations/BaseTest.java
@@ -1,0 +1,30 @@
+package annotations;
+
+import org.junit.After;
+import org.junit.Before;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import pom.MainPage;
+
+import java.time.Duration;
+
+public class BaseTest {
+
+    public WebDriver driver;
+
+    @Before
+    public void setUpChrome (){
+        ChromeOptions chromeOptions = new ChromeOptions();
+        chromeOptions.addArguments("--start-maximized");
+        driver = new ChromeDriver(chromeOptions);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
+        MainPage mainPage = new MainPage(driver);
+        mainPage.open();
+    }
+
+    @After
+    public void tearDown(){
+        driver.quit();
+    }
+}
