+++ src/test/java/login/LoginTest.java
@@ -5,28 +5,9 @@ import pom.LoginPage;
 import pom.MainPage;
 import pom.RegisterPage;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 
-import java.time.Duration;
-
-public class LoginTest {
-
-    private WebDriver driver;
-
-    @Before
-    public void setUpChrome(){
-        ChromeOptions chromeOptions = new ChromeOptions();
-        chromeOptions.addArguments("--start-maximized");
-        driver = new ChromeDriver(chromeOptions);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-        MainPage mainPage = new MainPage(driver);
-        mainPage.open();
-    }
+public class LoginTest extends annotations.BaseTest{
 
     @Test
     @DisplayName("Проверь вход через кнопку «Личный кабинет»")
@@ -80,10 +61,5 @@ public class LoginTest {
         loginPage.enterEmailAndPassword();
         loginPage.clickSignInButton();
         mainPage.checkOrderButton();
-    }
-
-    @After
-        public void tearDown(){
-        driver.quit();
     }
 }
