+++ src/test/java/login/LogoutTest.java
@@ -4,26 +4,9 @@ import pom.LoginPage;
 import pom.MainPage;
 import pom.ProfilePage;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 
-import java.time.Duration;
-
-public class LogoutTest {
-
-    private WebDriver driver;
-
-    @Before
-    public void setUpChrome(){
-        ChromeOptions chromeOptions = new ChromeOptions();
-        chromeOptions.addArguments("--start-maximized");
-        driver = new ChromeDriver(chromeOptions);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-    }
+public class LogoutTest extends annotations.BaseTest{
 
     @Test
     @DisplayName("Проверь выход по кнопке «Выйти» в личном кабинете.")
@@ -32,7 +15,6 @@ public class LogoutTest {
         LoginPage loginPage = new LoginPage(driver);
         ProfilePage profilePage = new ProfilePage(driver);
 
-        mainPage.open();
         mainPage.clickAccountButton();
         loginPage.enterEmailAndPassword();
         loginPage.clickSignInButton();
@@ -42,10 +24,5 @@ public class LogoutTest {
         mainPage.clickAccountButton();
         profilePage.clickLogoutButton();
         loginPage.checkRegistrationIsSuccessfully();
-    }
-
-    @After
-    public void tearDown(){
-        driver.quit();
     }
 }
