+++ src/test/java/registration/ShortPasswordErrorTest.java
@@ -1,30 +1,15 @@
 package registration;
 
+import com.github.javafaker.Faker;
+import io.qameta.allure.junit4.DisplayName;
+import org.junit.Test;
 import pom.LoginPage;
 import pom.MainPage;
 import pom.RegisterPage;
-import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import com.github.javafaker.Faker;
-import org.openqa.selenium.chrome.ChromeOptions;
 
-import java.time.Duration;
+public class ShortPasswordErrorTest extends annotations.BaseTest{
 
-public class ShortPasswordErrorTest {
-        private WebDriver driver;
         Faker faker = new Faker();
-
-        @Before
-        public void setUpChrome(){
-            ChromeOptions chromeOptions = new ChromeOptions();
-            chromeOptions.addArguments("--start-maximized");
-            driver = new ChromeDriver(chromeOptions);
-            driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-        }
 
         @Test
         @DisplayName("Проверь ошибку для некорректного пароля. Минимальный пароль — шесть символов.")
@@ -33,7 +18,6 @@ public class ShortPasswordErrorTest {
             LoginPage loginPage = new LoginPage(driver);
             RegisterPage registerPage = new RegisterPage(driver);
 
-            mainPage.open();
             mainPage.clickAccountButton();
             loginPage.clickRegisterButton();
             registerPage.inputName(faker.name().firstName());
@@ -41,10 +25,5 @@ public class ShortPasswordErrorTest {
             registerPage.inputPassword(faker.internet().password(3, 5));
             registerPage.clickFinallyRegisterButton();
             registerPage.checkShortPasswordError();
-        }
-
-        @After
-        public void tearDown(){
-            driver.quit();
         }
 }
